+++ core/src/test/java/dagger/internal/TestingModuleAdapter.java
@@ -24,24 +24,21 @@ import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
-import java.util.Map;
 import java.util.Set;
 import javax.inject.Provider;
 import javax.inject.Singleton;
 
 //TODO: Reduce the complexity of this and/or replace with a mock or fake.
 public class TestingModuleAdapter<M> extends ModuleAdapter<M> {
-  final Class<?> moduleClass;
-
-  public TestingModuleAdapter(Class<?> moduleClass, Module annotation) {
+  public TestingModuleAdapter(Class<M> moduleClass, Module annotation) {
     super(
+        moduleClass,
         injectableTypesToKeys(annotation.injects()),
         annotation.staticInjections(),
         annotation.overrides(),
         annotation.includes(),
         annotation.complete(),
         annotation.library());
-    this.moduleClass = moduleClass;
   }
 
   private static String[] injectableTypesToKeys(Class<?>[] injectableTypes) {
@@ -55,7 +52,7 @@ public class TestingModuleAdapter<M> extends ModuleAdapter<M> {
     return result;
   }
 
-  @Override public void getBindings(Map<String, Binding<?>> bindings) {
+  @Override public void getBindings(BindingsGroup bindings, M module) {
     for (Class<?> c = moduleClass; !c.equals(Object.class); c = c.getSuperclass()) {
       for (Method method : c.getDeclaredMethods()) {
         Provides provides = method.getAnnotation(Provides.class);
@@ -82,15 +79,15 @@ public class TestingModuleAdapter<M> extends ModuleAdapter<M> {
           String key = Keys.get(genericReturnType, method.getAnnotations(), method);
           switch (provides.type()) {
             case UNIQUE:
-              handleBindings(bindings, method, key, library);
+              handleBindings(bindings, module, method, key, library);
               break;
             case SET:
               String setKey = Keys.getSetKey(method.getGenericReturnType(),
                   method.getAnnotations(), method);
-              handleSetBindings(bindings, method, setKey, key, library);
+              handleSetBindings(bindings, module, method, setKey, key, library);
               break;
             case SET_VALUES:
-              handleSetBindings(bindings, method, key, key, library);
+              handleSetBindings(bindings, module, method, key, key, library);
               break;
             default:
               throw new AssertionError("Unknown @Provides type " + provides.type());
@@ -100,14 +97,17 @@ public class TestingModuleAdapter<M> extends ModuleAdapter<M> {
     }
   }
 
-  private void handleBindings(Map<String, Binding<?>> bindings, Method method, String key,
+  private void handleBindings(BindingsGroup bindings, M module, Method method, String key,
       boolean library) {
-    bindings.put(key, new ProviderMethodBinding<M>(method, key, module, library));
+    bindings.contributeProvidesBinding(key,
+        new ReflectiveProvidesBinding<M>(method, key, moduleClass.getName(), module, library));
   }
 
-  private void handleSetBindings(Map<String, Binding<?>> bindings, Method method, String setKey,
-      String providerKey, boolean library) {
-    SetBinding.<M>add(bindings, setKey, new ProviderMethodBinding<M>(method, providerKey, module, library));
+  private void handleSetBindings(BindingsGroup bindings, M module, Method method,
+      String setKey, String providerKey, boolean library) {
+    SetBinding.<M>add(bindings, setKey,
+        new ReflectiveProvidesBinding<M>(
+            method, providerKey, moduleClass.getName(), module, library));
   }
 
   @Override public M newModule() {
@@ -128,10 +128,14 @@ public class TestingModuleAdapter<M> extends ModuleAdapter<M> {
     }
   }
 
+  @Override public String toString() {
+    return "TestingModuleAdapter[" + this.moduleClass.getName() + "]";
+  }
+
   /**
    * Creates a TestingModuleAdapter or throws an {@code IllegalArgumentException}.
    */
-  public static <M> ModuleAdapter<M> create(Class<? extends M> moduleClass) {
+  public static <M> ModuleAdapter<M> create(Class<M> moduleClass) {
     Module annotation = moduleClass.getAnnotation(Module.class);
     if (annotation == null) {
       throw new IllegalArgumentException("No @Module on " + moduleClass.getName());
@@ -146,14 +150,14 @@ public class TestingModuleAdapter<M> extends ModuleAdapter<M> {
   /**
    * Invokes a method to provide a value. The method's parameters are injected.
    */
-  private final class ProviderMethodBinding<T> extends Binding<T> {
+  private static final class ReflectiveProvidesBinding<T> extends ProvidesBinding<T> {
     private Binding<?>[] parameters;
     private final Method method;
     private final Object instance;
 
-    public ProviderMethodBinding(Method method, String key, Object instance, boolean library) {
-      super(key, null, method.isAnnotationPresent(Singleton.class),
-          moduleClass.getName() + "." + method.getName() + "()");
+    public ReflectiveProvidesBinding(Method method, String key, String moduleClass,
+        Object instance, boolean library) {
+      super(key, method.isAnnotationPresent(Singleton.class), moduleClass, method.getName());
       this.method = method;
       this.instance = instance;
       method.setAccessible(true);
@@ -195,10 +199,6 @@ public class TestingModuleAdapter<M> extends ModuleAdapter<M> {
 
     @Override public void injectMembers(T t) {
       throw new AssertionError("Provides method bindings are not MembersInjectors");
-    }
-
-    @Override public String toString() {
-      return method.toString();
     }
   }
 }