+++ examples/android-activity-graphs/src/main/java/com/example/dagger/activitygraphs/ActivityModule.java
@@ -25,7 +25,7 @@ import javax.inject.Singleton;
 
 /**
  * This module represents objects which exist only for the scope of a single activity. We can
- * safely create singletons using the activity instance because ths entire object graph will only
+ * safely create singletons using the activity instance because the entire object graph will only
  * ever exist inside of that activity.
  */
 @Module(
@@ -33,7 +33,7 @@ import javax.inject.Singleton;
         HomeActivity.class,
         HomeFragment.class
     },
-    complete = false,
+    addsTo = AndroidModule.class,
     library = true
 )
 public class ActivityModule {