+++ CHANGELOG.md
@@ -1,6 +1,35 @@
 Change Log
 ==========
 
+Version 1.2.2 *(2014-07-21)*
+----------------------------
+
+ * Update JavaWriter to 2.5.0. This fixes incorrectly compressing fully-qualified class names
+   in child packages of `java.lang` (e.g., `java.lang.ref.WeakReference`).
+
+
+Version 1.2.1 *(2014-02-16)*
+----------------------------
+
+ * Restore Java 5 compatibility.
+ * New: Improve performance of `.plus()` with large volumes of set bindings.
+ * Fix: Do not mask underlying exception message from binding problems when constructing a graph.
+
+
+Version 1.2.0 *(2013-12-13)*
+----------------------------
+
+ * Numerous performance improvements in both the compiler and runtime.
+   * Use more efficient `String` concatenation.
+   * Module adapters are now stateless.
+   * Use read/write locks over global locks.
+   * Reflective constructor invocation is now cached with `Class.newInstance`.
+   * Avoid re-linking all bindings when calling `.plus()`.
+ * Set bindings are now unioned when calling `.plus()`.
+ * Fix: Tolerate missing type information during compilation by deferring writing
+   module adapters.
+
+
 Version 1.1.0 *(2013-08-05)*
 ----------------------------